# Small performance improvement: Switch urllib with simpler check