Reduces instantiation time from 9.005s to 8.480s on my system, should be functionally equivalent for our purposes, and is the fastest version I could think of as tested with pyperf:

```bash
$ pyperf timeit --setup 'from urllib.parse import urlparse' 'bool(urlparse("https://foobar.com/badabing.pdf").netloc)' 
.....................  
Mean +- std dev: 600 ns +- 6 ns

$ pyperf timeit '"https://foobar.com/badabing.pdf".startswith("http")' 
.....................  
Mean +- std dev: 40.0 ns +- 1.7 ns

$ pyperf timeit '"://" in "https://foobar.com/badabing.pdf"'
.....................  
Mean +- std dev: 13.5 ns +- 0.1 ns
```